int64 weights

+++ kahypar-resources/definitions.h
@@ -37,8 +37,8 @@
 namespace kahypar {
 using HypernodeID = uint32_t;
 using HyperedgeID = uint32_t;
-using HypernodeWeight = int32_t;
-using HyperedgeWeight = int32_t;
+using HypernodeWeight = int64_t;
+using HyperedgeWeight = int64_t;
 using PartitionID = int32_t;
 using Gain = HyperedgeWeight;
